my_flask_app/requirements.txt
flask
torch
faiss-cpu
transformers
datasets
scikit-learn
numpy
